Rename bidirectional → lang...
... and further tidying:

- organise scattered stdlib code into a module (including `nn`)
- `OperationDefinition` is now `TypedTerm`
- updates tests for new interface

TODO:

- Alias HashMap<Path, Operation>
- Make it easier to add stdlib items- currently verbose (see stdlib::nn)
