+++ catgrad-core/src/lib.rs
@@ -1,8 +1,12 @@
+// Core language
 pub mod category;
-pub mod check;
-pub mod nn;
-pub mod ssa;
-pub mod util;
+pub mod stdlib;
 
+// Shapechecking & Evaluation
+pub mod check;
 pub mod interpreter;
+pub mod ssa;
+
+// Utilities
 pub mod svg;
+pub mod util;
