+++ catgrad-core/src/stdlib/mod.rs
@@ -1,0 +1,5 @@
+//! Standard library of declarations and definitions for catgrad's `lang` category.
+pub mod ops;
+pub use ops::*;
+
+pub mod nn;
